Retry Enhancements
Adding in retry enhancements to WaitFor. There is an optional parameter added to the WaitFor along with a for loop to loop through the wait for a max amount of retries if specified.

If no retry is added added to the input, then it will continue as it did before.

There is also an optional string name added to the waitfor call for utilizaiton with the retry logic. This will allow for a more concise validation message on retry

XUnit additions were added back into this version as well